Option to allow for multiple systems.

+++ debclean.sh
@@ -2,7 +2,7 @@
 #
 # Change system back to state as defined in package list
 #
-BASE=debclean.baselist
+BASE=debclean.baselist-$(hostname)
 CUR=/tmp/debclean.curlist
 
 LOC=$( dirname ${BASH_SOURCE[0]})
@@ -21,19 +21,29 @@ case $1 in
    fi
 ;;
 "-c")
-   dpkg -l | grep "^.i" | awk '{print $2}' > $CUR
-   diff $BASE $CUR | grep '>' | cut -d' ' -f2 | xargs sudo apt-get -qq --purge -y remove
-   sudo apt-get -qq -y --purge autoremove
-   diff $BASE $CUR | grep '<' | cut -d' ' -f2 | xargs sudo apt-get -qq -y install
-   rm -f $CUR
+   if test -f $BASE
+   then
+      dpkg -l | grep "^.i" | awk '{print $2}' > $CUR
+      diff $BASE $CUR | grep '>' | cut -d' ' -f2 | xargs sudo apt-get -qq --purge -y remove
+      sudo apt-get -qq -y --purge autoremove
+      diff $BASE $CUR | grep '<' | cut -d' ' -f2 | xargs sudo apt-get -qq -y install
+      rm -f $CUR
+   else
+      echo "No base list is available for this system."
+   fi
 ;;
 "-l")
-   dpkg -l | grep "^.i" | awk '{print $2}' > $CUR
-   echo "Extra packages installed:"
-   diff $BASE $CUR | grep '>' | cut -d' ' -f2
-   echo "Packages removed:"
-   diff $BASE $CUR | grep '<' | cut -d' ' -f2
-   rm -f $CUR
+   if test -f $BASE
+   then
+      dpkg -l | grep "^.i" | awk '{print $2}' > $CUR
+      echo "Extra packages installed:"
+      diff $BASE $CUR | grep '>' | cut -d' ' -f2
+      echo "Packages removed:"
+      diff $BASE $CUR | grep '<' | cut -d' ' -f2
+      rm -f $CUR
+   else
+      echo "No base list is available for this system."
+   fi
 ;;
 *)
    echo "Usage: debclean.sh"
